clean up api assert methods
