+++ tests/Traits/Api/AssertApiPatchTrait.php
@@ -1,0 +1,36 @@
+<?php
+
+/*
+ * This file is part of the mangel.io project.
+ *
+ * (c) Florian Moser <git@famoser.ch>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace App\Tests\Traits\Api;
+
+use ApiPlatform\Core\Bridge\Symfony\Bundle\Test\Client;
+use Symfony\Component\HttpFoundation\Response as StatusCode;
+
+trait AssertApiPatchTrait
+{
+    private function assertApiPatchPayloadPersisted(Client $client, string $url, array $payload)
+    {
+        $response = $this->assertApiPatchOk($client, $url, $payload);
+        $this->assertJsonContains($payload);
+
+        return $response;
+    }
+
+    private function assertApiPatchOk(Client $client, string $url, array $payload)
+    {
+        return $this->assertApiPatchStatusCodeSame(StatusCode::HTTP_OK, $client, $url, $payload);
+    }
+
+    private function assertApiPatchStatusCodeSame(int $expectedCode, Client $client, string $url, array $payload)
+    {
+        return $this->assertApiStatusCodeSame('PATCH', $expectedCode, $client, $url, $payload);
+    }
+}
