+++ tests/Traits/Api/AssertApiResponseTrait.php
@@ -1,0 +1,68 @@
+<?php
+
+/*
+ * This file is part of the mangel.io project.
+ *
+ * (c) Florian Moser <git@famoser.ch>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace App\Tests\Traits\Api;
+
+use ApiPlatform\Core\Bridge\Symfony\Bundle\Test\Client;
+use ApiPlatform\Core\Bridge\Symfony\Bundle\Test\Response;
+
+trait AssertApiResponseTrait
+{
+    private function assertApiResponseFieldSubset(Response $response, string ...$expectedFields)
+    {
+        $content = $response->getContent();
+        $hydraPayload = json_decode($content, true);
+
+        $whitelist = array_merge(['@id', '@type'], $expectedFields);
+        sort($whitelist);
+
+        if ('hydra:Collection' === $hydraPayload['@type']) {
+            foreach ($hydraPayload['hydra:member'] as $member) {
+                $actualFields = array_keys($member);
+                sort($actualFields);
+
+                $this->assertArraySubset($actualFields, $whitelist);
+            }
+
+            return;
+        }
+
+        $this->fail('only collections support this assertion.');
+    }
+
+    private function assertApiResponseFileIsDownloadable(Client $client, Response $response, string $fileUrlProperty): ?string
+    {
+        $content = $response->getContent();
+        $hydraPayload = json_decode($content, true);
+
+        if ('hydra:Collection' === $hydraPayload['@type']) {
+            foreach ($hydraPayload['hydra:member'] as $member) {
+                $url = $member[$fileUrlProperty] ?? null;
+                if (!$url) {
+                    continue;
+                }
+
+                $response = $this->assertApiGetOk($client, $url);
+                $this->assertStringStartsWith('inline', $response->getHeaders()['content-disposition'][0]);
+
+                return $url;
+            }
+
+            $this->fail('no member has a the property '.$fileUrlProperty.' set, hence can not assert this url is valid.');
+
+            return null;
+        }
+
+        $this->fail('only collections support this assertion.');
+
+        return null;
+    }
+}
